Move generateChatTitle() to ChatbotComponent
